requirements.txt
opentldr
ipykernel
ipywidgets
python-dotenv
spacy
sentence-transformers
langchain
langchain-text-splitters
feedparser
beautifulsoup4
pymupdf
langchain_community
ollama
gpt4all
openai